requirements.txt
streamlit
openai-whisper
transformers
torch
sentence-transformers
spacy
nltk
bertopic
scikit-learn
moviepy
pydub
mlflow
